+++ app/Http/Controllers/ProfileController.js
@@ -1,4 +1,4 @@
-'use strict'
+'use strict';
 
 const Profile = use('App/Model/Profile');
 
@@ -13,30 +13,30 @@ class ProfileController {
     }
   }
 
-  * create(request, response) {
+  * create (request, response) {
     //
   }
 
-  * store(request, response) {
+  * store (request, response) {
     //
   }
 
-  * show(request, response) {
+  * show (request, response) {
     //
   }
 
-  * edit(request, response) {
+  * edit (request, response) {
     //
   }
 
-  * update(request, response) {
+  * update (request, response) {
     //
   }
 
-  * destroy(request, response) {
+  * destroy (request, response) {
     //
   }
 
 }
 
-module.exports = ProfileController
+module.exports = ProfileController;