+++ app/Http/Controllers/UserController.js
@@ -1,4 +1,4 @@
-'use strict'
+'use strict';
 
 const User = use('App/Model/User');
 
@@ -7,36 +7,36 @@ class UserController {
   * index (request, response) {
     try {
       const users = yield User.all();
-      response.status(200).json(users);
+      yield response.status(200).json(users);
     } catch (e) {
       response.send(e.message);
     }
   }
 
-  * create(request, response) {
+  * create (request, response) {
     //
   }
 
-  * store(request, response) {
+  * store (request, response) {
     //
   }
 
-  * show(request, response) {
+  * show (request, response) {
     //
   }
 
-  * edit(request, response) {
+  * edit (request, response) {
     //
   }
 
-  * update(request, response) {
+  * update (request, response) {
     //
   }
 
-  * destroy(request, response) {
+  * destroy (request, response) {
     //
   }
 
 }
 
-module.exports = UserController
+module.exports = UserController;