+++ config/database.js
@@ -1,4 +1,4 @@
-'use strict'
+'use strict';
 
 const Env = use('Env');
 const Helpers = use('Helpers');